requirements.txt
flask
pinecone
python-dotenv
pandas
openpyxl
tiktoken
sentence-transformers
torch